bug: changed deletePost to delete, still not working

+++ frontend/forum-app/src/app/posts/posts.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { PostsService } from '../posts.service';
+import { Storage } from '../storage';
 
 @Component({
   selector: 'app-posts',
@@ -31,10 +32,15 @@ export class PostsComponent implements OnInit {
 
   deletePost(id: string, title: string) { 
     console.log("Deleting post: " + title + "id: " + id);
-    this.service.deletePost(id).subscribe((response: any) => {
-      if (response.status == 200) {
-        this.snackbar.open("Post Deleted", "Dismiss", {duration: 1500 });
-      }
-    });
+    if (Storage.isLoggedIn) {
+      this.service.deletePost(id).subscribe((response: any) => {
+        if (response.status == 200) {
+          this.snackbar.open("Post Deleted", "Dismiss", {duration: 1500 });
+        }
+      });
+    }
+    else {
+      this.snackbar.open("You need to be logged in to delete posts", "Dismiss", {duration: 1500});
+    }
   }
 }
